Add features and Fix Bugs
I added the class Game to be a model of all possible sporting games. This allows me to add different sports to the smart alarm. The soccer class now checks if the game is in progress and will return the score. This commit also fixes the timezone support issue. The time difference between game time and user's local time now accounts for the user's timezone.